+++ tests/domain/DomainTests.cpp
@@ -5,7 +5,7 @@
 
 TEST(NodeSystem, PinValueAssignment) {
     of::domain::Pin inputPin;
-    inputPin.type = of::domain::PinType::FLOAT;
+    inputPin.type = of::domain::PinType::DOUBLE;
     inputPin.value = 42.0;
 
     ASSERT_TRUE(std::holds_alternative<double>(inputPin.value));
@@ -18,10 +18,10 @@ TEST(NodeSystem, NodeComponentStructure) {
 
     of::domain::Pin p1;
     p1.name = "A";
-    p1.type = of::domain::PinType::FLOAT;
+    p1.type = of::domain::PinType::DOUBLE;
     of::domain::Pin p2;
     p2.name = "B";
-    p2.type = of::domain::PinType::FLOAT;
+    p2.type = of::domain::PinType::DOUBLE;
 
     node.inputs.push_back(p1);
     node.inputs.push_back(p2);
